+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added a visibility eye icon in the Project tree between the object-type icon and the name (Gerber, Excellon, Geometry, CNC Job), following the light and dark tree colors
+
+### Fixed
+
+- Fixed Project tree range-selecting other layers when clicking a visibility eye while another row was already selected
+
 ## [1.9.0] - 2026-09-08
 
 ### Added